+++ tbears/command/command_util.py
@@ -28,6 +28,7 @@ class CommandUtil(object):
     def __init__(self, subparsers):
         self._add_init_parser(subparsers)
         self._add_samples_parser(subparsers)
+        self._add_genconf_parser(subparsers)
 
     @staticmethod
     def _add_init_parser(subparsers) -> None:
@@ -44,6 +45,11 @@ class CommandUtil(object):
                               help='Create two SCORE samples (standard_crowd_sale, standard_token)',
                               description='Create two SCORE samples (standard_crowd_sale, standard_token)')
 
+    @staticmethod
+    def _add_genconf_parser(subparser):
+        subparser.add_parser('genconf', help=f'Generate tbears config files. ({FN_CLI_CONF[2:]} and {FN_CLI_CONF[2:]})',
+                             description=f'Generate tbears config files. ({FN_CLI_CONF[2:]} and {FN_CLI_CONF[2:]})')
+
     def run(self, args):
         if not hasattr(self, args.command):
             print(f"Wrong command {args.command}")
@@ -59,6 +65,8 @@ class CommandUtil(object):
         """
         self._check_init(conf)
 
+        # initialize score project package. score class is set using main template.
+        # you can check main template at util/__init__/get_score_main_template method
         self.__initialize_project(project=conf['project'],
                                   score_class=conf['score_class'],
                                   contents_func=get_score_main_template)
@@ -69,12 +77,28 @@ class CommandUtil(object):
         """Generate two SCORE samples (standard_crowd_sale, standard_token)
         :param _conf: samples command configuration
         """
+        # initialize standard_token project package.
         self.__initialize_project(project="standard_token", score_class="StandardToken",
                                   contents_func=get_sample_token_contents)
+        # initialize standard_crowd_sale project package.
         self.__initialize_project(project="standard_crowd_sale", score_class="StandardCrowdSale",
                                   contents_func=get_sample_crowd_sale_contents)
 
         print(f"Made samples successfully")
+
+    def genconf(self, _conf: dict):
+        """Generate tbears config files. (tbears_server_config.json, tbears_cli_config.json)"""
+        result = []
+
+        if os.path.exists(FN_CLI_CONF) is False:
+            result.append(FN_CLI_CONF[2:])
+            write_file('./', FN_CLI_CONF, json.dumps(tbears_cli_config, indent=4))
+        if os.path.exists(FN_SERVER_CONF) is False:
+            result.append(FN_SERVER_CONF[2:])
+            write_file('./', FN_SERVER_CONF, json.dumps(tbears_server_config, indent=4))
+
+        if result:
+            print(f"Made {', '.join(result)} successfully")
 
     def check_command(self, command):
         return hasattr(self, command)
@@ -94,8 +118,12 @@ class CommandUtil(object):
         :param score_class: class name of SCORE.
         :param contents_func contents generator
         """
+        # make package.json data
         package_json_dict = get_package_json_dict(project, score_class)
         package_json_contents = json.dumps(package_json_dict, indent=4)
+
+        # when command is init, make score templete.
+        # when command is samples, make standard_crowd_sale or standard_token
         py_contents = contents_func(score_class)
 
         write_file(project, f"{project}.py", py_contents)