+++ tbears/libs/icx_signer.py
@@ -67,6 +67,7 @@ class IcxSigner:
 
         :return: base64-encoded string of recoverable signature data
         """
+        # 'msg_hash' argument must be 256bit bytes type data(made by hashlib.sha256() method)
         signature, recovery_id = self.sign_recoverable(msg_hash)
         recoverable_sig = bytes(bytearray(signature) + recovery_id.to_bytes(1, 'big'))
         return base64.b64encode(recoverable_sig)